% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleUpSet.R
\name{simpleUpSet}
\alias{simpleUpSet}
\title{Make simple UpSet plots}
\usage{
simpleUpSet(
  x,
  sets = NULL,
  sort_sets = TRUE,
  sort_intersect = TRUE,
  n_intersect = 20,
  min_size = 0,
  set_layers = default_set_layers(),
  intersect_layers = default_intersect_layers(),
  grid_layers = default_grid_layers(),
  highlight = NULL,
  annotations = list(),
  width = 0.75,
  height = 0.75,
  vjust_ylab = 0.8,
  stripe_colours = c("grey90", "white"),
  guides = "keep",
  top_left = NULL,
  ...,
  na.rm = TRUE
)
}
\arguments{
\item{x}{Input data frame}

\item{sets}{Character vector listing columns of x to plot}

\item{sort_sets}{Logical, if TRUE sort sets by size}

\item{sort_intersect}{Show intersections in decreasing order}

\item{n_intersect}{Maximum number of intersections to show}

\item{min_size}{Only show intersections larger than this value}

\item{set_layers}{List of \code{ggplot2} layers, scales and themes to define the
appearance of the sets panel. Can be obtained and extended using
\code{\link[=default_set_layers]{default_set_layers()}}}

\item{intersect_layers}{List of \code{ggplot2} layers, scales and themes to define
the appearance of the intersections panel. Can be obtained and extended
using \code{\link[=default_intersect_layers]{default_intersect_layers()}}}

\item{grid_layers}{List of \code{ggplot2} layers, scales & themes}

\item{highlight}{\code{\link[dplyr:case_when]{dplyr::case_when()}} statement defining all intersections to
highlight using \code{geom_intersect} and \code{scale_fill/colour_intersect}.
Will add a column named \code{highlight} which can be called from any geom}

\item{annotations}{list where each element is a list of ggplot2 layers.
Each element will be added as an upper annotation panel above the
intersections plot. All layer types (geom, scale, aes, stat, labs etc) can be
passed with the exception of facets.}

\item{width, height}{Proportional width and height of the intersection panel}

\item{vjust_ylab}{Used to nudge the y-axis labels closer to the axis}

\item{stripe_colours}{Colours for background stripes in the lower two panels.
For no stripes, set as NULL}

\item{guides}{Passed to \code{\link[patchwork:plot_layout]{patchwork::plot_layout()}}}

\item{top_left}{Optional ggplot object to show in the top left panel. Will
default to an empty ggplot object}

\item{...}{Not used}

\item{na.rm}{\code{NA} handling}
}
\value{
Object of class 'patchwork' containing multiple ggplot panels
}
\description{
Make simple UpSet plots using ggplot2 and patchwork
}
\details{
Taking a subset of columns from a data.frame, create an UpSet plot showing
all intersections as specified.
Columns chosen for the sets and intersections must contain logical values
or be strictly 0/1 values.

Additional columns can be used where appropriate for creating boxplots etc.
Internally, data objects will have the variables \code{set} and \code{intersect} which
can be referred to when passing custom aes() mappings to various layers.
Any additional layers passed using \code{annotations()} will have layers added
after an initial, internal call to \code{ggplot(data, aes(x = intersect))}.
}
\examples{
## Use the movies data from the package UpSetR
library(tidyverse)
theme_set(theme_bw())
f <- system.file("extdata", "movies.csv", package = "UpSetR")
movies <- read_delim(f, delim = ";") \%>\%
 mutate(
   ## Break the Release dates into decades
   Decade = cut(
     ReleaseDate, breaks = seq(1910, 2000, by = 10),
     labels = paste(seq(1911, 1991, by = 10), seq(1920, 2000, by = 10), sep = "-")
   ) \%>\% fct_rev()
 )
 sets <- c("Action", "Comedy", "Drama", "Thriller", "Romance")
## A basic UpSet plot
simpleUpSet(movies, sets)

## Fill intersects & sets by decade, showing fewer intersects
simpleUpSet(
  movies, sets, min_size = 20,
  geom_intersect = geom_bar(aes(fill = Decade)),
  geom_sets = geom_bar(aes(fill = Decade)),
  scale_fill_intersect = scale_fill_brewer(palette = "Paired"),
  scale_fill_sets = scale_fill_brewer(palette = "Paired"),
  theme_intersect = theme(
    legend.position = "inside",
    legend.position.inside = c(0.99, 0.99),
    legend.justification.inside = c(1, 1)
  ),
 theme_sets = theme(legend.position = "none")
)

## Add a simple boxplot as an additional upper panel
simpleUpSet(
  movies, sets, n_intersect = 10,
  annotations = list(geom_boxplot(aes(y = AvgRating))),
)

## Add a far more detailed upper plot
simpleUpSet(
  movies, sets, n_intersect = 10,
  annotations = list(
    list(
      aes(y = AvgRating),
      geom_jitter(aes(colour = Decade), height = 0, width = 0.3, alpha = 0.5),
      geom_violin(fill = NA, quantiles = 0.5, quantile.linetype = 1),
      scale_colour_brewer(palette = "Paired"),
      guides(colour = guide_legend(nrow = 2, reverse = TRUE))
    )
  ), guides = "collect"
) &
  theme(legend.position = "bottom")

## Modify set colours
set_cols <- c("darkred", "forestgreen", "blue", "blue", "blue")
names(set_cols) <- sets
simpleUpSet(
  movies, sets,
  geom_sets = geom_bar(aes(fill = set)),
  scale_fill_sets = scale_fill_manual(values = set_cols),
  theme_sets = theme(legend.position = "none")
)

}
