% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleUpSet.R
\name{simpleUpSet}
\alias{simpleUpSet}
\title{Make simple UpSet plots}
\usage{
simpleUpSet(
  x,
  sets = NULL,
  sort_sets = TRUE,
  sort_intersect = TRUE,
  n_intersect = 20,
  min_size = 0,
  fill_sets = NULL,
  fill_intersect = NULL,
  annotations = list(),
  set_labels = geom_text(hjust = 1.1, size = 4),
  intersect_labels = geom_text(vjust = -1, size = 4),
  intersect_points,
  intersect_segments,
  empty_intersect_points,
  set_label_fun = scales::comma,
  intersect_label_fun = set_label_fun,
  scale_x_sets,
  scale_y_intersect,
  scale_fill_sets,
  scale_fill_intersect,
  width = 0.75,
  height = 0.75,
  vjust_ylab = 0.8,
  stripe_colours = c("grey90", "white"),
  thm_intersect = list(),
  thm_sets = list(),
  thm_mat = list(),
  set_x_title = "Set Size",
  intersect_y_title = "Intersection Size",
  grid_x_title = "Intersection",
  guides = "keep",
  top_left = NULL,
  ...,
  na.rm = TRUE
)
}
\arguments{
\item{x}{Input data frame}

\item{sets}{Character vector listing columns of x to plot}

\item{sort_sets}{Logical, if TRUE sort sets by size}

\item{sort_intersect}{Show intersections in decreasing order}

\item{n_intersect}{Maximum number of intersections to show}

\item{min_size}{Only show intersections larger than this value}

\item{fill_sets, fill_intersect}{Additional column name in x to fill either
sets or intersections by. Each must be categorical}

\item{annotations}{list where each element is a list of ggplot2 layers.
Each element will be added as an upper annotation panel above the
intersections plot. All layer types (geom, scale, aes, stat, labs etc) can be
passed with the exception of facets. Elements will be added after an initial,
internal call to `ggplot(data, aes(x = intersect))`}

\item{set_labels, intersect_labels}{One of geom_text() or geom_label() with
arguments set as preferred. Use `NULL` to hide set or intersect labels}

\item{intersect_points, intersect_segments, empty_intersect_points}{Specifications of [geom_point()] and [geom_segment()] which can be used to
modify the defaults in the matrix panel}

\item{set_label_fun, intersect_label_fun}{Functions applied to set labels and
intersection labels}

\item{scale_x_sets, scale_y_intersect}{Optional scales to pass to these plots.
Can be used to over-ride default axis names, labelling and expansions}

\item{scale_fill_sets, scale_fill_intersect}{Discrete scales to use when
optionally filling bars by an additional column in 'x'}

\item{width, height}{Proportional width and height of the intersection panel}

\item{vjust_ylab}{Used to nudge the y-axis labels closer to the axis}

\item{stripe_colours}{Colours for background stripes in the lower two panels.
For no stripes, set as NULL}

\item{thm_intersect, thm_sets, thm_mat}{list objects containing
additional theme arguments for each panel}

\item{set_x_title, intersect_y_title, grid_x_title}{Axis titles for various
panels. Set to NULL to remove titles}

\item{guides}{Passed to [patchwork::plot_layout()]}

\item{top_left}{Optional ggplot object to show in the top left panel. Will
default to an empty ggplot object}

\item{...}{Not used}

\item{na.rm}{`NA` handling}
}
\value{
Object of class 'patchwork' containing multiple ggplot panels
}
\description{
Make simple UpSet plots using ggplot2 and patchwork
}
