% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleUpSet.R
\name{simpleUpSet}
\alias{simpleUpSet}
\title{Make simple UpSet plots}
\usage{
simpleUpSet(
  x,
  sets = NULL,
  sort_sets = TRUE,
  sort_intersect = TRUE,
  n_intersect = 20,
  min_size = 0,
  comma = TRUE,
  geom_intersect = geom_bar(),
  geom_sets = geom_bar(),
  geom_set_labels = geom_text(hjust = 1.1, size = 3.5),
  geom_intersect_labels = geom_text(vjust = -0.5, size = 3.5),
  grid_points = geom_point(size = 4, shape = 19),
  grid_segments = geom_segment(),
  empty_grid_points = geom_point(size = 4, shape = 19, colour = "grey70"),
  highlight = NULL,
  scale_x_sets = scale_x_reverse(name = "Set Size"),
  scale_y_intersect = scale_y_continuous(name = "Intersection Size"),
  scale_fill_sets = scale_fill_discrete(),
  scale_colour_sets = scale_colour_discrete(),
  scale_fill_intersect = scale_fill_discrete(),
  scale_colour_intersect = scale_colour_discrete(),
  scale_colour_grid = scale_colour_discrete(),
  scale_fill_grid = scale_fill_discrete(),
  annotations = list(),
  width = 0.75,
  height = 0.75,
  vjust_ylab = 0.8,
  stripe_colours = c("grey90", "white"),
  theme_intersect = NULL,
  theme_sets = NULL,
  theme_grid = NULL,
  grid_x_title = NULL,
  guides = "keep",
  top_left = NULL,
  ...,
  na.rm = TRUE
)
}
\arguments{
\item{x}{Input data frame}

\item{sets}{Character vector listing columns of x to plot}

\item{sort_sets}{Logical, if TRUE sort sets by size}

\item{sort_intersect}{Show intersections in decreasing order}

\item{n_intersect}{Maximum number of intersections to show}

\item{min_size}{Only show intersections larger than this value}

\item{comma}{logical(1) Add commas to intersection and set totals}

\item{geom_intersect, geom_sets}{Default geoms passed to intersections and
sets. Can only be \code{\link[ggplot2:geom_bar]{ggplot2::geom_bar()}}, but can be modified as needed}

\item{geom_set_labels, geom_intersect_labels}{One of \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}} or \code{\link[ggplot2:geom_text]{ggplot2::geom_label()}}
with arguments set as preferred. Use \code{NULL} to hide set or intersect labels}

\item{grid_points, grid_segments, empty_grid_points}{Specifications of \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} and \code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}}
which can be used to modify the defaults in the matrix panel intersection labels}

\item{highlight}{\code{\link[dplyr:case_when]{dplyr::case_when()}} statement defining all intersections to
highlight using \code{geom_intersect} and \code{scale_fill/colour_intersect}.
Will add a column named \code{highlight} which can be called from any geom}

\item{scale_x_sets, scale_y_intersect}{Optional scales to pass to these plots.
Can be used to over-ride default axis names, labelling and expansions}

\item{scale_fill_sets, scale_fill_intersect, scale_colour_sets, scale_colour_intersect}{Discrete scales to use when optionally filling or outlining bars by an
additional column in 'x'}

\item{scale_colour_grid, scale_fill_grid}{Colour scales passed to all points
and segments in the intersections matrix}

\item{annotations}{list where each element is a list of ggplot2 layers.
Each element will be added as an upper annotation panel above the
intersections plot. All layer types (geom, scale, aes, stat, labs etc) can be
passed with the exception of facets.}

\item{width, height}{Proportional width and height of the intersection panel}

\item{vjust_ylab}{Used to nudge the y-axis labels closer to the axis}

\item{stripe_colours}{Colours for background stripes in the lower two panels.
For no stripes, set as NULL}

\item{theme_intersect, theme_sets, theme_grid}{list objects containing
additional theme arguments for each panel}

\item{grid_x_title}{Optional axis title for the intersections matrix}

\item{guides}{Passed to \code{\link[patchwork:plot_layout]{patchwork::plot_layout()}}}

\item{top_left}{Optional ggplot object to show in the top left panel. Will
default to an empty ggplot object}

\item{...}{Not used}

\item{na.rm}{\code{NA} handling}
}
\value{
Object of class 'patchwork' containing multiple ggplot panels
}
\description{
Make simple UpSet plots using ggplot2 and patchwork
}
\details{
Taking a subset of columns from a data.frame, create an UpSet plot showing
all intersections as specified.
Columns chosen for the sets and intersections must contain logical values
or be strictly 0/1 values.

Additional columns can be used where appropriate for creating boxplots etc.
Internally, data objects will have the variables \code{set} and \code{intersect} which
can be referred to when passing custom aes() mappings to various layers.
Any additional layers passed using \code{annotations()} will have layers added
after an initial, internal call to \code{ggplot(data, aes(x = intersect))}.
}
\examples{
## Use the movies data from the package UpSetR
library(tidyverse)
theme_set(theme_bw())
f <- system.file("extdata", "movies.csv", package = "UpSetR")
movies <- read_delim(f, delim = ";") \%>\%
 mutate(
   ## Break the Release dates into decades
   Decade = cut(
     ReleaseDate, breaks = seq(1910, 2000, by = 10),
     labels = paste(seq(1911, 1991, by = 10), seq(1920, 2000, by = 10), sep = "-")
   ) \%>\% fct_rev()
 )
 sets <- c("Action", "Comedy", "Drama", "Thriller", "Romance")
## A basic UpSet plot
simpleUpSet(movies, sets)

## Fill intersects & sets by decade, showing fewer intersects
simpleUpSet(
  movies, sets, min_size = 20,
  geom_intersect = geom_bar(aes(fill = Decade)),
  geom_sets = geom_bar(aes(fill = Decade)),
  scale_fill_intersect = scale_fill_brewer(palette = "Paired"),
  scale_fill_sets = scale_fill_brewer(palette = "Paired"),
  theme_intersect = theme(
    legend.position = "inside",
    legend.position.inside = c(0.99, 0.99),
    legend.justification.inside = c(1, 1)
  ),
 theme_sets = theme(legend.position = "none")
)

## Add a simple boxplot as an additional upper panel
simpleUpSet(
  movies, sets, n_intersect = 10,
  annotations = list(geom_boxplot(aes(y = AvgRating))),
)

## Add a far more detailed upper plot
simpleUpSet(
  movies, sets, n_intersect = 10,
  annotations = list(
    list(
      aes(y = AvgRating),
      geom_jitter(aes(colour = Decade), height = 0, width = 0.3, alpha = 0.5),
      geom_violin(fill = NA, quantiles = 0.5, quantile.linetype = 1),
      scale_colour_brewer(palette = "Paired"),
      guides(colour = guide_legend(nrow = 2, reverse = TRUE))
    )
  ), guides = "collect"
) &
  theme(legend.position = "bottom")

## Modify set colours
set_cols <- c("darkred", "forestgreen", "blue", "blue", "blue")
names(set_cols) <- sets
simpleUpSet(
  movies, sets,
  geom_sets = geom_bar(aes(fill = set)),
  scale_fill_sets = scale_fill_manual(values = set_cols),
  theme_sets = theme(legend.position = "none")
)

}
